making provisioning an action against virtual template
